% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_template.R
\name{project_template}
\alias{project_template}
\title{GLA R Project Template}
\usage{
project_template(path, ...)
}
\arguments{
\item{path}{Path where project should be created}

\item{...}{Other parameters supplied by RStudio UI}
}
\value{
A new project session will begin
}
\description{
Function used by RStudio to create a project/package in the
GLA template
}
\details{
This function will be called when the user invokes the New Project
wizard using the project template defined in the template file
at: inst/rstudio/templates/project/project_template.dcf
}
